global: rename service/ to services/
