+++ tests/test_permissions.py
@@ -25,9 +25,9 @@ from invenio_records_permissions.policies.records import RecordPermissionPolicy
 
 from invenio_records_resources.resources import RecordResource, \
     RecordResourceConfig
-from invenio_records_resources.service import RecordService, \
+from invenio_records_resources.services import RecordService, \
     RecordServiceConfig
-from invenio_records_resources.service.errors import PermissionDeniedError
+from invenio_records_resources.services.errors import PermissionDeniedError
 
 HEADERS = {"content-type": "application/json", "accept": "application/json"}
